configHandler.py:  No longer depends on fileFinder
The case must now match for Morrowind.bsa, Tribunal.bsa, and Bloodmoon.bsa on case sensitive file systems.
The amount of effort to pretend a case sensitive file system isn't wasn't worth it here.

Note:
We still detect, and correct case errors for '.esp' and '.esm' files.
We even have a warning if two files with the same name, but different case are used on a case sensitive file system.
However, we make no guarantees which one will be used, and which will be ignored.